+++ src/test/java/tests/BaseTest.java
@@ -1,38 +1,38 @@
-package config;
-
-import static com.codeborne.selenide.Selenide.closeWebDriver;
-import static com.codeborne.selenide.logevents.SelenideLogger.addListener;
-import static helpers.AttachmentHelpers.attachAsText;
-import static helpers.AttachmentHelpers.attachPageSource;
-import static helpers.AttachmentHelpers.attachScreenshot;
-import static helpers.AttachmentHelpers.attachVideo;
-import static helpers.AttachmentHelpers.getConsoleLogs;
+package tests;
 
 import com.codeborne.selenide.Configuration;
-
+import config.ConfigHelper;
+import config.WebDriverConfig;
+import io.qameta.allure.selenide.AllureSelenide;
 import org.aeonbits.owner.ConfigFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.openqa.selenium.remote.DesiredCapabilities;
 
-import io.qameta.allure.selenide.AllureSelenide;
+import static com.codeborne.selenide.Selenide.closeWebDriver;
+import static com.codeborne.selenide.logevents.SelenideLogger.addListener;
+import static helpers.AttachmentHelpers.*;
 
 public class BaseTest {
 
     @BeforeAll
     static void setup() {
-        final WebDriverConfig config = ConfigFactory.create(WebDriverConfig.class, System.getProperties());
+        //final WebDriverConfig config = ConfigFactory.create(WebDriverConfig.class, System.getProperties());
         addListener("AllureSelenide", new AllureSelenide().screenshots(true).savePageSource(true));
+
         Configuration.startMaximized = true;
-        Configuration.browser = config.getBrowser();
-        Configuration.browserVersion = config.getBrowserVersion();
-        if (config.getSelenoidUrl() != null) {
-            Configuration.remote = config.getSelenoidUrl();
+        Configuration.browser = ConfigHelper.getBrowser();
+        Configuration.browserVersion = ConfigHelper.getBrowserVersion();
+
+        /*
+        if (ConfigHelper.getSelenoidUrl() != null) {
+            Configuration.remote = ConfigHelper.getSelenoidUrl();
             DesiredCapabilities capabilities = new DesiredCapabilities();
             capabilities.setCapability("enableVNC", true);
             capabilities.setCapability("enableVideo", true);
             Configuration.browserCapabilities = capabilities;
         }
+         */
     }
 
     @AfterEach
@@ -42,7 +42,7 @@ public class BaseTest {
         attachAsText("Browser console logs", getConsoleLogs());
         if (System.getProperty("video_storage") != null)
             attachVideo();
+
         closeWebDriver();
     }
 }
-
