+++ src/test/resources/local.properties
@@ -1,2 +1,2 @@
 browser=firefox
-browser.version=45.4.0
+browser.version=86.0
