+++ src/test/resources/remote.properties
@@ -1,3 +1,3 @@
-browser=chrome
-browser.version=89.0.4389.82
-selenoid.url=https://user1:1234@selenoid.autotests.cloud:4444/wd/hub/
+//browser=chrome
+//browser.version=89.0.4389.82
+//selenoid.url=https://user1:1234@selenoid.autotests.cloud:4444/wd/hub/
